+++ frontend/src/views/spell-editor/components/SpellIconSelector.vue
@@ -1,10 +1,16 @@
 <template>
-  <div v-if="icons" class="align-content-center">
+  <div
+    v-if="icons"
+    class="align-content-center"
+    id="spell-icon-view-port"
+    style="height: 90vh; overflow-y: scroll"
+  >
     <div
       @click="selectIcon(icon)"
       v-for="icon in icons"
       :key="icon"
       :class="'d-inline-block '"
+      :id="'spell-icon-' + icon"
       style="margin: 2px"
     >
       <span
@@ -17,8 +23,8 @@
 </template>
 
 <script>
-import {App}    from "../../../constants/app";
 import EqAssets from "../../../app/eq-assets/eq-assets";
+import util     from "util";
 
 export default {
   name: "SpellIconSelector",
@@ -36,6 +42,22 @@ export default {
   },
   created() {
     this.icons = EqAssets.getSpellIcons()
+
+    if (this.selectedIcon > 0) {
+      // we need 100ms delay because the videos haven't been rendered yet
+      setTimeout(() => {
+        console.time("[SpellIconSelector] scrollTo");
+        const container = document.getElementById("spell-icon-view-port");
+        const target    = document.getElementById(util.format("spell-icon-%s", this.selectedIcon))
+
+        // 230 is height of video to offset
+        if (container && target) {
+          container.scrollTop = container.scrollTop + target.getBoundingClientRect().top - 250;
+        }
+        console.timeEnd("[SpellIconSelector] scrollTo");
+      }, 1)
+    }
+
   },
   methods: {
     selectIcon(icon) {